redux implemented
