+++ src/components/CreateOrderForm.css
@@ -1,5 +1,5 @@
-.table-order-page {
-    background-color: rgba(0, 67, 50, 255);
+.add-order-heading{
+    color: rgba(0, 67, 50, 255);
 }
 
 .order-page-content {
@@ -8,9 +8,6 @@
     flex-direction: column;
     align-items: center;
     min-height: 70vh;
-    border-style: solid;
-    border-width: 1px;
-    border-color: rgba(192, 161, 89, 255);
 }
 
 .order-item {
@@ -22,8 +19,6 @@
     grid-template-columns: 300px 300px 300px;
     justify-items: center;
     align-items: center;
-    box-sizing: border-box;
-    border: 2px solid rgba(0, 67, 50, 255);
     grid-template-rows: 35px 50px;
 }
 
@@ -33,12 +28,13 @@
 }
 
 .quantity-buttons-wrapper, .add-item-button-wrapper{
-    background-color: white;
     height: 35px;
 }
 
 .quantity-buttons-wrapper{
-    width: 63px;
+    width: 65px;
+    display: flex;
+    justify-content: center;
 }
 
 .add-item-button-wrapper{
@@ -48,6 +44,7 @@
 .quantity-update-button, .add-item-button{
     padding: 10px;
     background: none;
+    color: rgba(192, 161, 89, 255);
 }
 
 .quantity-update-button{
@@ -67,30 +64,53 @@
     flex-direction: column;
     gap: 10px;
     margin: 40px;
-    padding: 40px;
-    color: rgba(192, 161, 89, 255);
+    padding: 10px;
+    color: rgba(0, 67, 50, 255);
 }
 
 .view-added-elements-wrapper:not(:empty) {
     border-style: solid;
     border-width: 1px;
-    border-color: rgba(192, 161, 89, 255);
+    border-color: rgba(0, 67, 50, 255);
   }
 
 .dropdown-label{
-    color: rgba(192, 161, 89, 255);
+    color: rgba(0, 67, 50, 255);
+    font-weight: 600;
+    font-size: x-large;
+    font-family: fangsong, serif;
 }
 
 .add-item-button{
     width: 104%;
-    padding: 10px;
     background-color: rgba(0, 67, 50, 255);
-    border-width: 1px;
-    border-color: rgba(192, 161, 89, 255);
-    color: rgba(192, 161, 89, 255);
 }
 
 .success-message{
-    color: rgba(192, 161, 89, 255);
+    color: rgba(0, 67, 50, 255);
     text-align: center;
   }
+
+.create-order-button{
+    width: 25%;
+    padding: 10px;
+}
+
+.quantity-container{
+    align-self: center;
+    color: rgba(192, 161, 89, 255);
+}
+
+.added-elements-quantity-buttons{
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    justify-items: center;
+    align-items: center;
+}
+
+.decrement-button, .increment-button{
+ border-radius: 70%;
+ height: 30px;
+ width: 30px;
+ font-size: larger;
+}
