+++ src/components/Layout.css
@@ -1,0 +1,3 @@
+.Layout-wrapper{
+    background-color: rgba(192, 161, 89, 255);
+}
