+++ src/pages/Home.css
@@ -1,8 +1,8 @@
 @import url('https://fonts.googleapis.com/css2?family=Playfair+Display+SC:wght@900&display=swap');
 
-.home {
+/* .home {
     background-color: rgba(192, 161, 89, 255);
-}
+} */
 
 .home-content {
     display: flex;
@@ -10,8 +10,12 @@
     align-items: center;
 }
 
-.welcome-message-container {
+.welcome-text-container {
     padding: 10px;
+}
+
+.select-table-text-container{
+    margin-top: 10px;
 }
 
 .table-container {
@@ -19,7 +23,7 @@
     grid-template-rows: 90px 90px 90px;
     grid-template-columns: 90px 90px 90px 90px;
     grid-gap: 20px;
-    padding: 40px 0px;
+    padding: 15px 0px 40px 0px;
 }
 
 .table {
@@ -32,3 +36,13 @@
     border-radius: 8px;
     text-decoration: none;
 }
+
+.homepage-text{
+    font-family: fangsong, serif;
+    font-size: larger;
+}
+
+/* select{
+    background-color: rgba(0, 67, 50, 1);
+    color: rgba(192, 161, 89, 255);
+} */
