+++ src/pages/Table.css
@@ -1,0 +1,3 @@
+.table-order-page {
+    display: flex;
+}
